// src/api/models/getMembershipEnergyMetricsPeriod.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Conluz API
 * Conluz is an API-driven application designed for the efficient management of an energy community,enabling the administration of community members and their corresponding supply points and the retrieval of consumption, production data.
 * OpenAPI spec version: 1.0.0
 */

export type GetMembershipEnergyMetricsPeriod = typeof GetMembershipEnergyMetricsPeriod[keyof typeof GetMembershipEnergyMetricsPeriod];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const GetMembershipEnergyMetricsPeriod = {
  LATEST_PUBLISHED_MONTH: 'LATEST_PUBLISHED_MONTH',
} as const;